remove smt package for questions and update refs
