+++ projects/question/src/main/java/org/batfish/question/SmtBoundedLengthQuestionPlugin.java
@@ -1,4 +1,4 @@
-package org.batfish.question.smt;
+package org.batfish.question;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import org.batfish.common.Answerer;
@@ -6,9 +6,8 @@ import org.batfish.common.plugin.IBatfish;
 import org.batfish.datamodel.answers.AnswerElement;
 import org.batfish.datamodel.questions.Question;
 import org.batfish.datamodel.questions.smt.HeaderLocationQuestion;
-import org.batfish.question.QuestionPlugin;
 
-public class BoundedLengthQuestionPlugin extends QuestionPlugin {
+public class SmtBoundedLengthQuestionPlugin extends QuestionPlugin {
 
   public static class BoundedLengthAnswerer extends Answerer {
