+++ projects/question/src/main/java/org/batfish/question/SmtMultipathConsistencyQuestionPlugin.java
@@ -1,13 +1,12 @@
-package org.batfish.question.smt;
+package org.batfish.question;
 
 import org.batfish.common.Answerer;
 import org.batfish.common.plugin.IBatfish;
 import org.batfish.datamodel.answers.AnswerElement;
 import org.batfish.datamodel.questions.Question;
 import org.batfish.datamodel.questions.smt.HeaderLocationQuestion;
-import org.batfish.question.QuestionPlugin;
 
-public class MultipathConsistencyQuestionPlugin extends QuestionPlugin {
+public class SmtMultipathConsistencyQuestionPlugin extends QuestionPlugin {
 
   public static class MulipathConsistencyAnswerer extends Answerer {
