+++ projects/question/src/main/java/org/batfish/question/SmtReachabilityQuestionPlugin.java
@@ -1,13 +1,12 @@
-package org.batfish.question.smt;
+package org.batfish.question;
 
 import org.batfish.common.Answerer;
 import org.batfish.common.plugin.IBatfish;
 import org.batfish.datamodel.answers.AnswerElement;
 import org.batfish.datamodel.questions.Question;
 import org.batfish.datamodel.questions.smt.HeaderLocationQuestion;
-import org.batfish.question.QuestionPlugin;
 
-public class ReachabilityQuestionPlugin extends QuestionPlugin {
+public class SmtReachabilityQuestionPlugin extends QuestionPlugin {
 
   @Override
   protected Answerer createAnswerer(Question question, IBatfish batfish) {
