+++ tests/java-smt/aggregation2-forwarding.ref
@@ -62,7 +62,7 @@
     }
   ],
   "question" : {
-    "class" : "org.batfish.question.smt.ForwardingQuestionPlugin$ForwardingQuestion",
+    "class" : "org.batfish.question.SmtForwardingQuestionPlugin$ForwardingQuestion",
     "differential" : false,
     "dstIps" : [
       "42.42.42.1"
