+++ tests/java-smt/bgp1-bounded-len2.ref
@@ -66,7 +66,7 @@
     }
   ],
   "question" : {
-    "class" : "org.batfish.question.smt.BoundedLengthQuestionPlugin$BoundedLengthQuestion",
+    "class" : "org.batfish.question.SmtBoundedLengthQuestionPlugin$BoundedLengthQuestion",
     "bound" : 0,
     "differential" : false,
     "equivalence" : false,
