+++ tests/java-smt/campus1-local.ref
@@ -33,7 +33,7 @@
     }
   ],
   "question" : {
-    "class" : "org.batfish.question.smt.LocalConsistencyQuestionPlugin$LocalConsistencyQuestion",
+    "class" : "org.batfish.question.SmtLocalConsistencyQuestionPlugin$LocalConsistencyQuestion",
     "differential" : false,
     "equivalence" : false,
     "failures" : 0,
