+++ tests/java-smt/campus1-reachability.ref
@@ -3203,7 +3203,7 @@
     }
   ],
   "question" : {
-    "class" : "org.batfish.question.smt.ReachabilityQuestionPlugin$ReachabilityQuestion",
+    "class" : "org.batfish.question.SmtReachabilityQuestionPlugin$ReachabilityQuestion",
     "differential" : false,
     "equivalence" : false,
     "failures" : 0,
