+++ tests/java-smt/env3-local.ref
@@ -161,7 +161,7 @@
     }
   ],
   "question" : {
-    "class" : "org.batfish.question.smt.LocalConsistencyQuestionPlugin$LocalConsistencyQuestion",
+    "class" : "org.batfish.question.SmtLocalConsistencyQuestionPlugin$LocalConsistencyQuestion",
     "differential" : false,
     "equivalence" : false,
     "failures" : 0,
