+++ tests/java-smt/ospf3-bounded-len.ref
@@ -10,7 +10,7 @@
     }
   ],
   "question" : {
-    "class" : "org.batfish.question.smt.BoundedLengthQuestionPlugin$BoundedLengthQuestion",
+    "class" : "org.batfish.question.SmtBoundedLengthQuestionPlugin$BoundedLengthQuestion",
     "bound" : 2,
     "differential" : false,
     "failures" : 0,
