+++ tests/java-smt/ospf3-bounded-len2.ref
@@ -98,7 +98,7 @@
     }
   ],
   "question" : {
-    "class" : "org.batfish.question.smt.BoundedLengthQuestionPlugin$BoundedLengthQuestion",
+    "class" : "org.batfish.question.SmtBoundedLengthQuestionPlugin$BoundedLengthQuestion",
     "bound" : 1,
     "differential" : false,
     "equivalence" : false,
