+++ tests/java-smt/ospf3-multipath.ref
@@ -10,7 +10,7 @@
     }
   ],
   "question" : {
-    "class" : "org.batfish.question.smt.MultipathConsistencyQuestionPlugin$MultipathConsistencyQuestion",
+    "class" : "org.batfish.question.SmtMultipathConsistencyQuestionPlugin$MultipathConsistencyQuestion",
     "differential" : false,
     "failures" : 0,
     "finalIfaceRegex" : "Loopback.*",
