+++ tests/java-smt/overflow1-forwarding.ref
@@ -82,7 +82,7 @@
     }
   ],
   "question" : {
-    "class" : "org.batfish.question.smt.ForwardingQuestionPlugin$ForwardingQuestion",
+    "class" : "org.batfish.question.SmtForwardingQuestionPlugin$ForwardingQuestion",
     "differential" : false,
     "dstIps" : [
       "41.41.41.1"
